会员后台category分类更新 2016-03-18

+++ resources/views/member/category/index.blade.php
@@ -26,9 +26,14 @@
                 <td>{{ $data->pid }}</td>
                 <td>{{ $data->created_at }}</td>
                 <td>
-                    <a href="/member/category/{{ $data->id }}" class="list_btn">查看</a>
-                    <a href="/member/category/{{ $data->id }}/edit" class="list_btn">编辑</a>
-                    <a href="/member/category/{{ $data->id }}/destroy" class="list_btn">删除</a>
+                    @if($curr=='')
+                        <a href="/member/category/{{ $data->id }}" class="list_btn">查看</a>
+                        <a href="/member/category/{{ $data->id }}/edit" class="list_btn">编辑</a>
+                        <a href="/member/category/{{ $data->id }}/destroy" class="list_btn">删除</a>
+                    @elseif($curr=='trash')
+                        <a href="/member/category/{{ $data->id }}/restore" class="list_btn">还原</a>
+                        <a href="/member/category/{{ $data->id }}/forceDelete" class="list_btn">销毁记录</a>
+                    @endif
                 </td>
             </tr>
             @endforeach
