+++ resources/views/member/persondemand/index.blade.php
@@ -26,10 +26,15 @@
                 <td>{{ $data->uname }}</td>
                 <td>{{ $data->created_at }}</td>
                 <td>
-                    <a href="/member/persondemand/{{ $data->id }}/pre" class="list_btn">预览</a>
-                    <a href="/member/persondemand/{{ $data->id }}" class="list_btn">查看</a>
-                    <a href="/member/persondemand/{{ $data->id }}/edit" class="list_btn">编辑</a>
-                    <a href="/member/persondemand/{{ $data->id }}/destroy" class="list_btn">删除</a>
+                    @if($curr=='')
+                        <a href="/member/persondemand/{{ $data->id }}/pre" class="list_btn">预览</a>
+                        <a href="/member/persondemand/{{ $data->id }}" class="list_btn">查看</a>
+                        <a href="/member/persondemand/{{ $data->id }}/edit" class="list_btn">编辑</a>
+                        <a href="/member/persondemand/{{ $data->id }}/destroy" class="list_btn">删除</a>
+                    @elseif($curr=='trash')
+                        <a href="/member/persondemand/{{ $data->id }}/restore" class="list_btn">还原</a>
+                        <a href="/member/persondemand/{{ $data->id }}/forceDelete" class="list_btn">销毁记录</a>
+                    @endif
                 </td>
             </tr>
             @endforeach
